Address comments - II
