+++ keras_nlp/models/deberta/deberta_backbone.py
@@ -12,88 +12,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""DeBERTa backbone models."""
+"""DeBERTa backbone model."""
 
 import tensorflow as tf
 from tensorflow import keras
 
-from keras_nlp.models.deberta.deberta_encoder import DebertaEncoder
+from keras_nlp.models.deberta.disentangled_attention_encoder import (
+    DisentangledAttentionEncoder,
+)
+from keras_nlp.models.deberta.relative_embedding import RelativeEmbedding
 from keras_nlp.utils.python_utils import classproperty
 
 
 def deberta_kernel_initializer(stddev=0.02):
     return keras.initializers.TruncatedNormal(stddev=stddev)
-
-
-class RelativeEmbedding(keras.layers.Layer):
-    """Relative embedding layer.
-
-    This layer initializes an embedding matrix (of shape
-    `(2 * batch_size, hidden_dim)`) for relative position encoding. It then
-    applies layer normalization on the embedding matrix and returns the relative
-    embedding matrix. This is in accordance with the original DeBERTa V3
-    implementation.
-
-    Args:
-        hidden_dim: int. The size of the dense embedding.
-        bucket_size: int. The size of the relative position buckets.
-        layer_norm_epsilon: float. Epsilon value to initialize the layer
-            normalization layer.
-        kernel_initializer: string or `keras.initializers` initializer,
-            defaults to "glorot_uniform". The kernel initializer for
-            the dense embedding.
-    """
-
-    def __init__(
-        self,
-        hidden_dim,
-        bucket_size,
-        layer_norm_epsilon=1e-05,
-        kernel_initializer="glorot_uniform",
-        name=None,
-        **kwargs,
-    ):
-        super().__init__(name=name, **kwargs)
-
-        self.hidden_dim = hidden_dim
-        self.bucket_size = bucket_size
-        self.layer_norm_epsilon = layer_norm_epsilon
-        self.kernel_initializer = keras.initializers.get(kernel_initializer)
-
-        self.rel_embeddings = self.add_weight(
-            shape=(self.bucket_size * 2, self.hidden_dim),
-            initializer=self.kernel_initializer,
-            name="rel_embedding",
-        )
-        self.layer_norm = keras.layers.LayerNormalization(
-            epsilon=layer_norm_epsilon, name="rel_embeddings_layer_norm"
-        )
-
-    def call(self, inputs):
-        batch_size = tf.shape(inputs)[0]
-
-        rel_embeddings = self.rel_embeddings[tf.newaxis, :]
-        rel_embeddings = self.layer_norm(rel_embeddings)
-
-        # Repeat `rel_embeddings` along axis = 0 `batch_size` times. The
-        # resultant shape is `(batch_size, bucket_size * 2, hidden_dim)`.
-        rel_embeddings = tf.repeat(rel_embeddings, repeats=batch_size, axis=0)
-
-        return rel_embeddings
-
-    def get_config(self):
-        config = super().get_config()
-        config.update(
-            {
-                "hidden_dim": self.hidden_dim,
-                "bucket_size": self.bucket_size,
-                "layer_norm_epsilon": self.layer_norm_epsilon,
-                "kernel_initializer": keras.initializers.serialize(
-                    self.kernel_initializer
-                ),
-            }
-        )
-        return config
 
 
 @keras.utils.register_keras_serializable(package="keras_nlp")
@@ -133,8 +65,7 @@ class DebertaBackbone(keras.Model):
     Example usage:
     ```python
     input_data = {
-        "token_ids": tf.random.uniform(
-            shape=(1, 12), dtype=tf.int64, maxval=128100),
+        "token_ids": tf.ones(shape=(1, 12), dtype=tf.int64),
         "padding_mask": tf.constant(
             [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0], shape=(1, 12)),
     }
@@ -206,7 +137,7 @@ class DebertaBackbone(keras.Model):
 
         # Apply successive DeBERTa encoder blocks.
         for i in range(num_layers):
-            x = DebertaEncoder(
+            x = DisentangledAttentionEncoder(
                 num_heads=num_heads,
                 intermediate_dim=intermediate_dim,
                 max_position_embeddings=max_sequence_length,
@@ -216,7 +147,7 @@ class DebertaBackbone(keras.Model):
                     x, approximate=False
                 ),
                 kernel_initializer=deberta_kernel_initializer(),
-                name=f"deberta_encoder_layer_{i}",
+                name=f"disentangled_attention_encoder_layer_{i}",
             )(
                 x,
                 rel_embeddings=rel_embeddings,
@@ -262,7 +193,7 @@ class DebertaBackbone(keras.Model):
 
     @classproperty
     def presets(cls):
-        raise NotImplementedError
+        return {}
 
     @classmethod
     def from_preset(
